test(fidelity-gate): split hardening tests into their own file
test_test_gate_fidelity_triage.py crossed the 300-line bloat gate a
second time after the Tier-3 fix's regression tests landed. Split the
resolved-obligation-narrowing tests and the Tier-3 hardening tests
(malformed screen data, escaping symlinks) into
test_test_gate_fidelity_triage_hardening.py, sharing the fixture
helpers back via a relative import (precedent:
test_verifiers_design_screens.py / test_context_cost_hook_registration_additive.py).
